% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbatteries-package.R
\docType{package}
\name{rbatteries-package}
\alias{rbatteries}
\alias{rbatteries-package}
\title{rbatteries: Help write R with all the batteries included}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CGMossa/rbatteries}
  \item Report bugs at \url{https://github.com/CGMossa/rbatteries/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
\keyword{internal}
