% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_generic.R
\name{resample}
\alias{resample}
\alias{resample.default}
\alias{resample.data.frame}
\title{Resample}
\usage{
resample(x, ...)

\method{resample}{default}(x, ...)

\method{resample}{data.frame}(x, ...)
}
\arguments{
\item{x}{Object to resample}

\item{...}{Arguments passed on to \code{base::sample.int}
\describe{
  \item{n}{a positive number, the number of items to choose from.  See
    \sQuote{Details.}}
  \item{size}{a non-negative integer giving the number of items to choose.}
  \item{replace}{should sampling be with replacement?}
  \item{prob}{a vector of probability weights for obtaining the elements
    of the vector being sampled.}
  \item{useHash}{\code{\link{logical}} indicating if the hash-version of
    the algorithm should be used.  Can only be used for \code{replace =
      FALSE}, \code{prob = NULL}, and \code{size <= n/2}, and really
    should be used for large \code{n}, as \code{useHash=FALSE} will use
    memory proportional to \code{n}.}
}}
}
\description{
Resample
}
