% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{quartiles}
\alias{quartiles}
\alias{deciles}
\alias{percentiles}
\title{Tiles
\code{quartiles} returns the quartiles, \code{deciles} return the deciles, and \code{percentiles} returns the percentiles with a resolution of \code{0.01}.}
\usage{
quartiles(x)

deciles(x)

percentiles(x)
}
\arguments{
\item{x}{numeric vector whose sample quantiles are wanted, or an
    object of a class for which a method has been defined (see also
    \sQuote{details}). \code{\link{NA}} and \code{NaN} values are not
    allowed in numeric vectors unless \code{na.rm} is \code{TRUE}.}
}
\value{
Returns the quartiles of \code{x}
}
\description{
Tiles
\code{quartiles} returns the quartiles, \code{deciles} return the deciles, and \code{percentiles} returns the percentiles with a resolution of \code{0.01}.
}
\section{Functions}{
\itemize{
\item \code{deciles}: The deciles of \code{x}

\item \code{percentiles}: The percentiles of \code{x}
}}

