% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{quartiles}
\alias{quartiles}
\alias{deciles}
\alias{percentiles}
\title{*Tiles}
\usage{
quartiles(x, na.rm = FALSE, names = TRUE, type = 7, ...)

deciles(x, na.rm = FALSE, names = TRUE, type = 7, ...)

percentiles(x, na.rm = FALSE, names = TRUE, type = 7, ...)
}
\arguments{
\item{x}{numeric vector whose sample quantiles are wanted, or an
    object of a class for which a method has been defined (see also
    \sQuote{details}). \code{\link{NA}} and \code{NaN} values are not
    allowed in numeric vectors unless \code{na.rm} is \code{TRUE}.}

\item{na.rm}{logical; if true, any \code{\link{NA}} and \code{NaN}'s
    are removed from \code{x} before the quantiles are computed.}

\item{names}{logical; if true, the result has a \code{\link{names}}
    attribute.  Set to \code{FALSE} for speedup with many \code{probs}.}

\item{type}{an integer between 1 and 9 selecting one of the
    nine quantile algorithms detailed below to be used.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Returns the quartiles of \code{x}.
}
\description{
\code{quartiles} returns the quartiles, \code{deciles} return the deciles, and \code{percentiles} returns the percentiles with a resolution of \code{0.01}.
}
\section{Functions}{
\itemize{
\item \code{deciles}: Returns the deciles of \code{x}.

\item \code{percentiles}: Returns the percentiles of \code{x}.
}}

\examples{
x <- rnorm(10)
quartiles(x)
deciles(x)
percentiles(x)
}
\seealso{
\itemize{
\item \code{\link[stats:quantile]{stats::quantile()}}
}
}
\concept{stats::quantile}
